+++ app/src/Navigators/Main.tsx
@@ -8,6 +8,8 @@ import { ExampleContainer, ListContainer } from '@/Containers'
 import Entypo from 'react-native-vector-icons/Entypo'
 import { ListType } from '@/Containers/ListContainer'
 import { useTheme } from '@/Hooks'
+import ProfileContainer from '@/NativeBase/Containers/ProfileContainer'
+import VerticalStackContainer from '@/NativeBase/Containers/VerticalStackContainer'
 
 const Tab = createBottomTabNavigator()
 
@@ -38,10 +40,21 @@ const MainNavigator = () => {
         }}
       />
       <Tab.Screen
-        name="Example"
-        component={ExampleContainer}
+        name="Container"
+        component={VerticalStackContainer}
         options={{
-          tabBarIcon: () => <Entypo color={Colors.text} name="cog" size={24} />,
+          tabBarIcon: () => (
+            <Entypo color={Colors.text} name="folder" size={24} />
+          ),
+        }}
+      />
+      <Tab.Screen
+        name="Profile"
+        component={ProfileContainer}
+        options={{
+          tabBarIcon: () => (
+            <Entypo color={Colors.text} name="user" size={24} />
+          ),
         }}
       />
     </Tab.Navigator>