+++ app/src/Navigators/Application.tsx
@@ -16,7 +16,6 @@ import {
   EventDetailContainer,
   StartupContainer,
   ProjectDetailContainer,
-  ProjectSearchContainer,
   EventSearchContainer,
   WelcomeContainer,
 } from '@/Containers'
@@ -25,6 +24,7 @@ import MainNavigator from './Main'
 import { navigationRef } from './utils'
 import ProjectScope from '@/Components/Project/ProjectScope'
 import NavigationHeader from '@/NativeBase/Components/NavigationHeader'
+import ProjectSearchContainer from '@/NativeBase/Containers/ProjectSearchContainer'
 import StaTheme from '@/NativeBase/Theme/StaTheme'
 import { ThemeState } from '@/Store/Theme'
 
@@ -151,7 +151,7 @@ const ApplicationNavigator = () => {
             component={ProjectSearchContainer}
             options={{
               ...stackScreenDefaultOptions,
-              title: 'Project Search',
+              title: 'Search',
             }}
           />